# [json] Use S.R.I vectors in JsonReaderHelper