… src/libraries/System.Text.Json/src/System/Text/Json/Reader/JsonReaderHelper.sri.cs