@radekdoulik 2023-02-07
Now to measure the CLR perf.