@radekdoulik 2023-02-07
```
> dotnet run --base ..\..\benchmarks\micro\artifacts\before --diff ..\..\benchmarks\micro\artifacts\after --threshold 2%
summary:
better: 170, geomean: 1.120
worse: 15, geomean: 1.080
total diff: 185
```
| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.32 |           174.67 |           230.27 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: False, TestCas |      1.18 |            91.47 |           107.96 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.13 |          1964.42 |          2211.09 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ImmutableSortedDictionary<String,  |      1.09 |         64071.49 |         69845.51 |         |
| System.Text.Json.Serialization.Tests.ReadJson<LargeStructWithProperties>.Deseria |      1.09 |          1618.14 |          1760.57 |         |
| System.Text.Json.Serialization.Tests.ReadJson<SimpleStructWithProperties>.Deseri |      1.09 |           302.07 |           327.76 |         |
| System.Text.Json.Serialization.Tests.ReadJson<SimpleStructWithProperties>.Deseri |      1.06 |           356.91 |           377.47 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ImmutableSortedDictionary<String,  |      1.05 |         65186.69 |         68711.29 |         |
| System.Text.Json.Tests.Perf_Depth.ReadSpanEmptyLoop(Depth: 65)                   |      1.04 |          2604.85 |          2705.15 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: True, TestCase |      1.04 |          3612.22 |          3745.75 |         |
| System.Text.Json.Serialization.Tests.ReadJson<LargeStructWithProperties>.Deseria |      1.04 |          1096.43 |          1136.79 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.04 |           155.63 |           161.24 |         |
| System.Text.Json.Serialization.Tests.ReadJson<SimpleStructWithProperties>.Deseri |      1.03 |           348.85 |           360.75 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ImmutableSortedDictionary<String,  |      1.03 |         67303.70 |         69273.53 |         |
| System.Text.Json.Serialization.Tests.ReadJson<LargeStructWithProperties>.Deseria |      1.02 |          1119.73 |          1146.20 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Text.Json.Reader.Tests.Perf_Base64.ReadBase64EncodedByteArray_NoEscaping( |      1.42 |           106.42 |            74.98 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.35 |         54171.11 |         40269.90 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.34 |         81990.53 |         61079.60 |         |
| System.Text.Json.Tests.Perf_Segment.ReadSingleSegmentSequenceByN(numberOfBytes:  |      1.33 |        838665.13 |        629626.25 |         |
| System.Text.Json.Tests.Perf_Segment.ReadSingleSegmentSequenceByN(numberOfBytes:  |      1.33 |        801958.13 |        602781.25 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequence(segmentSize: 8192,  |      1.31 |        843846.71 |        645042.00 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.30 |          2130.97 |          1633.98 |         |
| System.Text.Json.Tests.Perf_Segment.ReadSingleSegmentSequence(TestCase: Json400K |      1.30 |        789941.88 |        606954.81 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: False, TestCas |      1.30 |         93536.98 |         72216.62 |         |
| System.Text.Json.Tests.Perf_Segment.ReadSingleSegmentSequenceByN(numberOfBytes:  |      1.29 |         79237.91 |         61536.18 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequence(segmentSize: 8192,  |      1.28 |         85788.83 |         66994.20 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.28 |         53131.10 |         41496.18 |         |
| System.Text.Json.Tests.Perf_Segment.ReadSingleSegmentSequence(TestCase: Json40KB |      1.28 |         78482.95 |         61379.17 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequenceUsingSpan(segmentSiz |      1.28 |        825731.58 |        646651.56 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequence(segmentSize: 4096,  |      1.27 |        841121.38 |        662677.86 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequence(segmentSize: 4096,  |      1.26 |         86886.03 |         68862.44 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: False, TestC |      1.26 |         78381.39 |         62192.98 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.25 |         54474.93 |         43674.11 |         |
| System.Text.Json.Tests.Perf_Segment.ReadSingleSegmentSequenceByN(numberOfBytes:  |      1.25 |         77304.46 |         62027.01 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.25 |          2003.03 |          1607.78 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequenceUsingSpan(segmentSiz |      1.24 |         76811.70 |         61760.95 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: True, TestCase |      1.24 |         64054.52 |         51556.49 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: False, TestC |      1.22 |          3309.63 |          2707.95 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.22 |           100.17 |            82.27 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.21 |          1257.69 |          1037.76 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.20 |         79941.10 |         66622.79 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequenceUsingSpan(segmentSiz |      1.20 |          7608.40 |          6349.20 |         |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromStr |      1.20 |         28761.76 |         24011.45 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.19 |          8920.87 |          7471.39 | several?|
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequenceUsingSpan(segmentSiz |      1.19 |        839708.88 |        704922.83 |         |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromRea |      1.19 |         35938.47 |         30224.77 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromUtf8Byte |      1.19 |         21762.25 |         18352.61 |         |
| System.Text.Json.Serialization.Tests.ReadJson<MyEventsListerViewModel>.Deseriali |      1.18 |        418062.50 |        352830.00 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Dictionary<String, String>>.Deseri |      1.18 |         21659.53 |         18336.88 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: False, TestC |      1.18 |          1928.63 |          1634.17 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.18 |          7796.48 |          6612.16 |         |
| System.Text.Json.Serialization.Tests.ReadJson<MyEventsListerViewModel>.Deseriali |      1.17 |        426903.17 |        363564.82 |         |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromRea |      1.17 |         36337.92 |         30999.38 |         |
| System.Text.Json.Serialization.Tests.ReadJson<MyEventsListerViewModel>.Deseriali |      1.17 |        312598.63 |        267386.44 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequenceUsingSpan(segmentSiz |      1.17 |         82573.17 |         70753.89 |         |
| System.Text.Json.Tests.Perf_Segment.ReadSingleSegmentSequenceByN(numberOfBytes:  |      1.17 |          7236.55 |          6210.89 |         |
| System.Text.Json.Serialization.Tests.ReadJson<MyEventsListerViewModel>.Deseriali |      1.16 |        303100.66 |        262059.70 |         |
| System.Text.Json.Serialization.Tests.ReadJson<MyEventsListerViewModel>.Deseriali |      1.16 |        315906.71 |        273245.65 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromStream(M |      1.15 |         22244.77 |         19279.93 |         |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromUtf |      1.15 |         25776.66 |         22347.23 |         |
| System.Text.Json.Serialization.Tests.ReadJson<MyEventsListerViewModel>.Deseriali |      1.15 |        304603.62 |        264193.54 |         |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromUtf |      1.15 |         26104.20 |         22644.35 |         |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromStr |      1.15 |         26893.72 |         23334.47 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: False, TestC |      1.15 |         11422.54 |          9913.06 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: True, TestCase |      1.15 |            96.44 |            83.71 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.15 |          1026.31 |           894.37 |         |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromStr |      1.15 |         26812.74 |         23399.71 |         |
| System.Text.Json.Tests.Perf_Segment.ReadSingleSegmentSequenceByN(numberOfBytes:  |      1.14 |          7155.73 |          6257.28 |         |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromStr |      1.14 |         28460.07 |         24904.35 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.14 |          3467.05 |          3042.88 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromReader(M |      1.14 |         24705.74 |         21684.42 |         |
| System.Text.Json.Serialization.Tests.ReadJson<MyEventsListerViewModel>.Deseriali |      1.14 |        348169.97 |        306564.01 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.13 |           622.57 |           549.56 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Dictionary<String, String>>.Deseri |      1.13 |         21393.31 |         18923.10 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: False, TestC |      1.13 |            91.91 |            81.30 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: False, TestC |      1.13 |          6876.16 |          6093.51 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromString(M |      1.13 |         21209.04 |         18797.52 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Nullable<DateTimeOffset>>.Deserial |      1.13 |           262.20 |           232.95 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromStream(M |      1.12 |         21805.44 |         19415.31 |         |
| System.Text.Json.Serialization.Tests.ReadJson<MyEventsListerViewModel>.Deseriali |      1.12 |        347375.62 |        309636.34 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromUtf8Byte |      1.12 |         21000.20 |         18738.51 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Dictionary<String, String>>.Deseri |      1.12 |         17321.19 |         15503.53 |         |
| System.Text.Json.Reader.Tests.Perf_Base64.ReadBase64EncodedByteArray_HeavyEscapi |      1.12 |            83.81 |            75.08 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.12 |          3282.79 |          2942.93 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromStream(M |      1.11 |         33555.98 |         30116.36 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromString(M |      1.11 |         21569.83 |         19359.33 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: False, TestCas |      1.11 |         13050.39 |         11716.77 |         |
| System.Text.Json.Serialization.Tests.ReadJson<HashSet<String>>.DeserializeFromSt |      1.11 |         11415.93 |         10264.64 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromReader(M |      1.11 |         38098.55 |         34270.56 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.11 |          7266.47 |          6538.15 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.11 |          4522.99 |          4071.31 |         |
| System.Text.Json.Reader.Tests.Perf_Base64.ReadBase64EncodedByteArray_NoEscaping( |      1.11 |           258.01 |           232.29 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: False, TestCas |      1.11 |          7969.61 |          7179.94 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.11 |          4685.97 |          4222.20 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromReader(M |      1.11 |         24335.70 |         21931.73 |         |
| System.Text.Json.Serialization.Tests.ReadMissingAndCaseInsensitive<Location>.Mis |      1.11 |           822.24 |           741.03 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromStream(M |      1.11 |         33820.33 |         30480.44 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequence(segmentSize: 4096,  |      1.11 |          7279.79 |          6574.63 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.11 |          3156.51 |          2855.79 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequence(segmentSize: 8192,  |      1.10 |          6921.42 |          6269.15 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.10 |           741.51 |           671.91 |         |
| System.Text.Json.Tests.Perf_Depth.ReadSpanEmptyLoop(Depth: 64)                   |      1.10 |          2546.23 |          2310.43 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.10 |            86.22 |            78.28 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.10 |          7419.63 |          6736.33 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromReader(M |      1.10 |         37544.22 |         34115.83 |         |
| System.Text.Json.Reader.Tests.Perf_Base64.ReadBase64EncodedByteArray_HeavyEscapi |      1.10 |           256.44 |           233.38 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.10 |          7132.67 |          6500.95 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.10 |         12569.76 |         11463.01 |         |
| System.Text.Json.Tests.Perf_Depth.ReadSpanEmptyLoop(Depth: 512)                  |      1.10 |         21750.94 |         19836.66 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromString(M |      1.10 |         32965.73 |         30067.96 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: True, TestCase |      1.10 |          5620.48 |          5126.45 |         |
| System.Text.Json.Serialization.Tests.ReadJson<HashSet<String>>.DeserializeFromUt |      1.10 |         10241.48 |          9343.25 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.10 |           905.02 |           826.01 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: False, TestC |      1.09 |           886.41 |           809.60 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: False, TestC |      1.09 |          9146.05 |          8356.03 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromString(M |      1.09 |         33490.36 |         30654.80 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.09 |          3086.99 |          2826.70 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromUtf8Byte |      1.09 |         32250.46 |         29657.21 |         |
| System.Text.Json.Tests.Perf_Segment.ReadMultiSegmentSequenceUsingSpan(segmentSiz |      1.09 |          6951.29 |          6396.26 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Nullable<DateTimeOffset>>.Deserial |      1.09 |           134.13 |           123.50 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Location>.DeserializeFromReader(Mo |      1.08 |          1737.49 |          1602.14 |         |
| System.Text.Json.Serialization.Tests.ReadJson<LoginViewModel>.DeserializeFromRea |      1.08 |           672.36 |           621.36 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Dictionary<String, String>>.Deseri |      1.08 |         16654.39 |         15391.53 |         |
| System.Text.Json.Serialization.Tests.ReadJson<HashSet<String>>.DeserializeFromUt |      1.08 |          9963.63 |          9220.99 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: True, TestCase |      1.08 |           741.15 |           686.11 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Nullable<DateTimeOffset>>.Deserial |      1.08 |           253.47 |           235.11 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.08 |           609.54 |           565.90 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromUtf8Byte |      1.08 |         32994.09 |         30646.49 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.08 |          6809.89 |          6330.47 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Nullable<DateTimeOffset>>.Deserial |      1.07 |           180.33 |           168.02 |         |
| System.Text.Json.Serialization.Tests.ReadJson<BinaryData>.DeserializeFromReader( |      1.07 |           637.58 |           594.60 |         |
…